refactor: VoucherType domain 레이어로 이동
