+++ src/main/java/com/programmers/springbootbasic/domain/voucher/PercentDiscountVoucher.java
@@ -1,7 +1,5 @@
 package com.programmers.springbootbasic.domain.voucher;
 
-import com.programmers.springbootbasic.service.VoucherType;
-
 import java.util.UUID;
 
 public class PercentDiscountVoucher extends Voucher {
