+++ src/main/java/com/programmers/springbootbasic/domain/voucher/Voucher.java
@@ -1,7 +1,5 @@
 package com.programmers.springbootbasic.domain.voucher;
 
-import com.programmers.springbootbasic.service.VoucherType;
-
 import java.time.LocalDateTime;
 import java.util.UUID;
